fix message error
