+++ src/main/java/com/qpark/controller/HomeController.java
@@ -1,0 +1,43 @@
+package com.qpark.controller;
+
+import com.qpark.DatabaseConnection;
+import com.qpark.dao.ParkingAreaDAO;
+import com.qpark.model.ParkingArea;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.io.IOException;
+import java.sql.Connection;
+import java.sql.SQLException;
+import java.util.List;
+
+@WebServlet(name = "HomeController", value = "")
+public class HomeController extends HttpServlet {
+    private ParkingAreaDAO parkingAreaDAO;
+
+    public void init() throws ServletException {
+        super.init();
+
+        try {
+            Connection connection = DatabaseConnection.getConnection();
+            parkingAreaDAO = new ParkingAreaDAO(connection);
+        } catch (SQLException e) {
+            e.printStackTrace();
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        try {
+            List<ParkingArea> parkingAreasList = parkingAreaDAO.findAll();
+            request.setAttribute("parkingAreasList", parkingAreasList);
+            request.getRequestDispatcher("/index.jsp").forward(request, response);
+        } catch (SQLException ex) {
+            throw new ServletException(ex);
+        }
+    }
+}
