+++ src/main/java/com/qpark/controller/ParkingSlotsController.java
@@ -1,5 +1,8 @@
 package com.qpark.controller;
 
+import com.qpark.DatabaseConnection;
+import com.qpark.dao.ParkingAreaDAO;
+import com.qpark.model.ParkingArea;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
@@ -7,10 +10,55 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
+import java.sql.Connection;
+import java.sql.SQLException;
 
 @WebServlet(name = "ParkingSlotsController", value = "/parking_slot")
 public class ParkingSlotsController extends HttpServlet {
+
+    private ParkingAreaDAO parkingAreaDAO;
+
+    public void init() throws ServletException {
+        super.init();
+
+        try {
+            Connection connection = DatabaseConnection.getConnection();
+            parkingAreaDAO = new ParkingAreaDAO(connection);
+        } catch (SQLException e) {
+            e.printStackTrace();
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        try {
+            bookParkingSlot(request, response);
+        } catch (SQLException ex) {
+            throw new ServletException(ex);
+        }
+    }
+
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        try {
+            viewParkingArea(request, response);
+
+        } catch (SQLException ex) {
+            throw new ServletException(ex);
+        }
+    }
+
+    private void viewParkingArea(HttpServletRequest request, HttpServletResponse response) throws SQLException, ServletException, IOException {
+        int parkingAreaId = Integer.parseInt(request.getParameter("id"));
+        ParkingArea parkingArea = parkingAreaDAO.findById(parkingAreaId);
+        request.setAttribute("parkingArea", parkingArea);
         request.getRequestDispatcher("/views/parking_slot.jsp").forward(request, response);
+    }
+
+    private void bookParkingSlot(HttpServletRequest request, HttpServletResponse response) throws SQLException, ServletException, IOException {
+        int parkingAreaId = Integer.parseInt(request.getParameter("id"));
+        ParkingArea parkingArea = parkingAreaDAO.findById(1);
+        request.setAttribute("parkingArea", parkingArea);
+        request.getRequestDispatcher("/index.jsp").forward(request, response);
     }
 }
